requirements.txt
python-dotenv==0.20.0
paramiko==2.11.0